scripts/requirements.txt
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
idna==3.4
polling2==0.5.0
python-dotenv==1.0.0
requests==2.28.2
requests-oauthlib==1.3.1
urllib3==1.26.14
